--- src/editor/Components/README.md
@@ -1,7 +1,0 @@
-## Elements
-
-Each element can be displayed as:
-
-  - Element itself (non-draggable, changing it will change all elements)
-  - Add menu and dragging
-  - Inside component
