+++ src/editor/Nodes/Box.tsx
@@ -1,9 +1,7 @@
-import { BoxNode, ComponentView, DragDirection, InputNode, Node } from '@src/interfaces'
+import { BoxNode, ComponentView, Node } from '@src/interfaces'
 import state from '@state'
-import DragCorners from '@src/editor/Components/DragCorners'
 import * as React from 'react'
 import styled, { css } from 'styled-components'
-import Component from '@src/editor/Components/Component'
 
 const selectComponent = (component: Node, parent: Node) => e => {
   if (e.currentTarget === e.target) {
@@ -78,15 +76,14 @@ const editBox = (component: Node) => e => {
 }
 
 interface BoxProps {
-  component: InputNode
+  component: BoxNode
   parent: Node
 }
 
-const Input = styled.input`
+const Boxxy = styled.div`
   transition: all 0.3s;
   position: relative;
   display: grid;
-  outline: none;
   opacity: ${({ parent }) => (state.ui.editingBoxNode && state.ui.editingBoxNode === parent ? 0.4 : 1)};
   grid-template-columns: ${({ component }: BoxProps) => component.columns.map(col => col.value + col.unit).join(' ')};
   grid-template-rows: ${({ component }: BoxProps) => component.rows.map(col => col.value + col.unit).join(' ')};
@@ -146,78 +143,14 @@ const Input = styled.input`
         `
       : ''};
 `
-interface BorderProps {
-  col: number
-  row: number
-}
-
-const Border = styled.div`
-  grid-column: ${({ col }: BorderProps) => `${col} / ${col + 1}`};
-  grid-row: ${({ row }: BorderProps) => `${row} / ${row + 1}`};
-
-  border: #565656 dashed 1px;
-  user-select: none;
-  z-index: 999999;
-`
-const stylesForSelected = (component: Node) => {
-  if (state.ui.selectedNode !== component || !state.ui.draggingNodePosition) {
-    return null
-  }
-
-  return {
-    transition: 'none',
-    zIndex: 999999,
-    pointerEvents: 'none',
-    opacity: '0.75',
-    transform: `translateX(${state.ui.draggingNodePosition.x}px) translateY(${state.ui.draggingNodePosition.y}px)`,
-  }
-}
-
-const changeGridSize = (rowIndex: number, colIndex: number) => () => {
-  const direction = state.ui.expandingNode.direction
-  const position = state.ui.expandingNode.node.position
-  const columnPositive = [DragDirection.E, DragDirection.SE, DragDirection.NE].includes(direction)
-  const columnNegative = [DragDirection.W, DragDirection.SW, DragDirection.NW].includes(direction)
-  const rowPositive = [DragDirection.S, DragDirection.SW, DragDirection.SE].includes(direction)
-  const rowNegative = [DragDirection.N, DragDirection.NE, DragDirection.NW].includes(direction)
-  if (columnPositive && colIndex + 2 > position.columnStart) {
-    position.columnEnd = colIndex + 2
-  }
-  if (columnNegative && colIndex + 1 < position.columnEnd) {
-    position.columnStart = colIndex + 1
-  }
-  if (rowPositive && rowIndex + 2 > position.rowStart) {
-    position.rowEnd = rowIndex + 2
-  }
-  if (rowNegative && rowIndex + 1 < position.rowEnd) {
-    position.rowStart = rowIndex + 1
-  }
-}
 
 const BoxComponent = ({ component, parent }: BoxProps) => (
-  <label>
-    {component.label}
-    <Input
-      parent={parent}
-      component={component}
-      onMouseDown={selectComponent(component, parent)}
-      onDoubleClick={editBox(component)}
-      style={stylesForSelected(component)}
-    />
-    <DragCorners component={component} parent={parent} />
-    {state.ui.expandingNode &&
-      state.ui.expandingNode.parent === component &&
-      component.rows.map((_, rowIndex) =>
-        component.columns.map((_, colIndex) => (
-          <Border
-            key={`${colIndex}_${rowIndex}`}
-            row={rowIndex + 1}
-            col={colIndex + 1}
-            onMouseOver={changeGridSize(rowIndex, colIndex)}
-          />
-        )),
-      )}
-  </label>
+  <Boxxy
+    parent={parent}
+    component={component}
+    onMouseDown={selectComponent(component, parent)}
+    onDoubleClick={editBox(component)}
+  />
 )
 
 export default BoxComponent
