+++ src/editor/Settings/colorList.ts
